Create README.md
Rainfall-Prediction
In this project, I will be implementing a predictive model on Rain Dataset to predict whether or not it will rain tomorrow in Australia.The Dataset contains about 10 years of daily weather observations of different locations in Australia. By the end of this project, you will be able to build a predictive model.